fix: expose types in export map
When using `nodenext` resolution in typescript, the compiler will not
fall back to top-level `types` and will instead expect there to be a
sibling `d.ts` file or a `types` entry in the export.

This means currently errorstacks fails builds in modern typescript
setups.

This fix introduces a `types` entry for the export, which points at the
same definition file as the top-level `types`.